Перевод приложения на Typescript (часть 4)
