+++ src/index.tsx
@@ -2,27 +2,17 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./components/App/App";
-import { rootReducer } from "./services/reducers/rootReducer";
-import { configureStore } from "@reduxjs/toolkit";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
+import { rootStore } from "./utils/store";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 
-const preloadedState = {};
-
-const store = configureStore({
-  reducer: rootReducer,
-  middleware: (getDefaultMiddleware) => getDefaultMiddleware(),
-  devTools: process.env.NODE_ENV !== "production",
-  preloadedState,
-});
-
 root.render(
   <React.StrictMode>
-    <Provider store={store}>
+    <Provider store={rootStore}>
       <BrowserRouter>
         <App />
       </BrowserRouter>
