--- src/services/logService.js
@@ -1,3 +1,0 @@
-export const logError = (error) => {
-  console.error("Ошибка:", error);
-};
