+++ src/services/logService.ts
@@ -1,0 +1,3 @@
+export const logError = (error: string | object) => {
+  console.error("Ошибка:", error);
+};
