+++ src/utils/store.ts
@@ -1,0 +1,22 @@
+import { rootReducer } from "../services/reducers/rootReducer";
+import { configureStore } from "@reduxjs/toolkit";
+import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
+
+const preloadedState = {};
+
+export const rootStore = configureStore({
+  reducer: rootReducer,
+  middleware: (getDefaultMiddleware) => getDefaultMiddleware(),
+  devTools: process.env.NODE_ENV !== "production",
+  preloadedState,
+});
+
+// https://redux.js.org/usage/usage-with-typescript
+// Infer the `RootState` and `AppDispatch` types from the store itself
+type RootState = ReturnType<typeof rootStore.getState>;
+// Inferred type: {posts: PostsState, comments: CommentsState, users: UsersState}
+type AppDispatch = typeof rootStore.dispatch;
+
+// Use throughout your app instead of plain `useDispatch` and `useSelector`
+export const useAppDispatch: () => AppDispatch = useDispatch;
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
